Decoupling ZooKeeper from ElasticJob
