+++ elasticjob-cloud/elasticjob-cloud-common/pom.xml
@@ -59,7 +59,7 @@
         </dependency>
         <dependency>
             <groupId>org.apache.shardingsphere.elasticjob</groupId>
-            <artifactId>elasticjob-registry-center</artifactId>
+            <artifactId>elasticjob-registry-center-zookeeper-curator</artifactId>
             <version>${project.parent.version}</version>
         </dependency>
         <dependency>
