+++ elasticjob-infra/elasticjob-registry-center/elasticjob-registry-center-api/src/main/java/org/apache/shardingsphere/elasticjob/reg/exception/IgnoredExceptionProvider.java
@@ -1,0 +1,16 @@
+package org.apache.shardingsphere.elasticjob.reg.exception;
+
+import java.util.Collection;
+
+/**
+ * Ignored exception provider.
+ */
+public interface IgnoredExceptionProvider {
+    
+    /**
+     * Get ignored exceptions.
+     *
+     * @return ignored exceptions
+     */
+    Collection<Class<? extends Throwable>> getIgnoredExceptions();
+}
