+++ elasticjob-infra/elasticjob-registry-center/elasticjob-registry-center-api/src/main/java/org/apache/shardingsphere/elasticjob/reg/exception/RegExceptionHandler.java
@@ -1,0 +1,62 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *  
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.shardingsphere.elasticjob.reg.exception;
+
+import lombok.AccessLevel;
+import lombok.NoArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+
+import java.util.Collection;
+import java.util.LinkedList;
+import java.util.ServiceLoader;
+
+/**
+ * Registry center exception handler.
+ */
+@Slf4j
+@NoArgsConstructor(access = AccessLevel.PRIVATE)
+public final class RegExceptionHandler {
+    
+    private static final Collection<Class<? extends Throwable>> IGNORED_EXCEPTIONS = new LinkedList<>();
+    
+    static {
+        ServiceLoader.load(IgnoredExceptionProvider.class).forEach(each -> IGNORED_EXCEPTIONS.addAll(each.getIgnoredExceptions()));
+    }
+    
+    /**
+     * Handle exception.
+     * 
+     * @param cause exception to be handled
+     */
+    public static void handleException(final Exception cause) {
+        if (null == cause) {
+            return;
+        }
+        if (isIgnoredException(cause) || null != cause.getCause() && isIgnoredException(cause.getCause())) {
+            log.debug("Elastic job: ignored exception for: {}", cause.getMessage());
+        } else if (cause instanceof InterruptedException) {
+            Thread.currentThread().interrupt();
+        } else {
+            throw new RegException(cause);
+        }
+    }
+    
+    private static boolean isIgnoredException(final Throwable cause) {
+        return IGNORED_EXCEPTIONS.stream().anyMatch(each -> each.isInstance(cause));
+    }
+}
