+++ elasticjob-infra/elasticjob-registry-center/elasticjob-regitry-center-provider/elasticjob-registry-center-zookeeper-curator/src/main/java/org/apache/shardingsphere/elasticjob/reg/zookeeper/exception/ZookeeperCuratorIgnoredExceptionProvider.java
@@ -1,0 +1,35 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.shardingsphere.elasticjob.reg.zookeeper.exception;
+
+import org.apache.shardingsphere.elasticjob.reg.exception.IgnoredExceptionProvider;
+import org.apache.zookeeper.KeeperException;
+
+import java.util.Arrays;
+import java.util.Collection;
+
+/**
+ * Zookeeper curator ignored exception provider.
+ */
+public final class ZookeeperCuratorIgnoredExceptionProvider implements IgnoredExceptionProvider {
+
+    @Override
+    public Collection<Class<? extends Throwable>> getIgnoredExceptions() {
+        return Arrays.asList(KeeperException.ConnectionLossException.class, KeeperException.NoNodeException.class, KeeperException.NodeExistsException.class);
+    }
+}
