+++ elasticjob-lite/elasticjob-lite-core/pom.xml
@@ -38,7 +38,7 @@
         </dependency>
         <dependency>
             <groupId>org.apache.shardingsphere.elasticjob</groupId>
-            <artifactId>elasticjob-registry-center</artifactId>
+            <artifactId>elasticjob-registry-center-api</artifactId>
             <version>${project.parent.version}</version>
         </dependency>
         <dependency>
